Changed to double

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,13 +8,13 @@
 int main(void)
 {
 	int i;
-	unsigned long long int fib1 = 1;
-	unsigned long long int fib2 = 2;
-	unsigned long long int future;
+	double fib1 = 1;
+	double fib2 = 2;
+	double future;
 
 	for (i = 0; i < 98; i++)
 	{
-		printf("%llu", fib1);
+		printf("%g", fib1);
 
 		future = fib1 + fib2;
 		fib1 = fib2;
